refactor: use vw_job_board
