+++ web/src/routes/(app)/account/jobs/+page.svelte
@@ -48,19 +48,19 @@
 				</thead>
 				<tbody class="divide-y divide-slate-200 bg-white">
 					{#each data.jobs as job}
-						<tr class="text-sm text-slate-700" class:bg-blue-50={job.is_verified}>
+						<tr class="text-sm text-slate-700" class:bg-sky-50={job.is_verified}>
 							<td class="table-cell py-4 pl-4 pr-3 text-base font-medium text-slate-900 lg:pl-6">
 								<div class="flex items-center space-x-4">
 									<div
-										class="flex h-10 w-10 items-center justify-center rounded-full bg-indigo-100"
+										class="flex h-10 w-10 items-center justify-center rounded-full bg-orange-100"
 										alt="Company Intial"
 									>
-										{job.company[0] || '?'}
+										{job.company_name[0] || '?'}
 									</div>
 									<div>
 										<div class="flex items-center space-x-2">
 											<span>
-												{job.company}
+												{job.company_name}
 											</span>
 											{#if job.is_verified}
 												<svg
@@ -85,16 +85,23 @@
 								</div>
 							</td>
 							<td class="w-full max-w-0 py-4 px-3 text-base text-slate-900 sm:w-auto sm:max-w-none">
-								{job.recruiter}
+								{job.recruiter_name}
 								<dl class="text-xs sm:text-sm">
 									<dt class="sr-only">Recruiter Email</dt>
 									<dd class="mt-1 truncate">{job.recruiter_email}</dd>
 								</dl>
 							</td>
 							<td class="px-3 py-4 text-sm">
-								<div class="space-y-1">
+								<div
+									class="flex items-center justify-end space-x-2 space-y-0 lg:flex-col lg:items-start lg:justify-start lg:space-x-0 lg:space-y-2"
+								>
 									{#if job.company_website}
-										<div class="flex items-center space-x-2">
+										<a
+											href={job.company_website}
+											target="_blank"
+											rel="noopener noreferrer"
+											class="flex items-center space-x-2 hover:text-blue-500"
+										>
 											<svg
 												xmlns="http://www.w3.org/2000/svg"
 												fill="none"
@@ -109,11 +116,18 @@
 													d="M12 21a9.004 9.004 0 008.716-6.747M12 21a9.004 9.004 0 01-8.716-6.747M12 21c2.485 0 4.5-4.03 4.5-9S14.485 3 12 3m0 18c-2.485 0-4.5-4.03-4.5-9S9.515 3 12 3m0 0a8.997 8.997 0 017.843 4.582M12 3a8.997 8.997 0 00-7.843 4.582m15.686 0A11.953 11.953 0 0112 10.5c-2.998 0-5.74-1.1-7.843-2.918m15.686 0A8.959 8.959 0 0121 12c0 .778-.099 1.533-.284 2.253m0 0A17.919 17.919 0 0112 16.5c-3.162 0-6.133-.815-8.716-2.247m0 0A9.015 9.015 0 013 12c0-1.605.42-3.113 1.157-4.418"
 												/>
 											</svg>
-											<span> Company Website </span>
-										</div>
+											<span class="hidden hover:underline lg:block">
+												{new URL(job.company_website).hostname}
+											</span>
+										</a>
 									{/if}
 									{#if job.job_description_url}
-										<div class="flex items-center space-x-2">
+										<a
+											class="flex items-center space-x-2 hover:text-blue-500"
+											href={job.job_description_url}
+											target="_blank"
+											rel="noopener noreferrer"
+										>
 											<svg
 												xmlns="http://www.w3.org/2000/svg"
 												fill="none"
@@ -128,8 +142,10 @@
 													d="M20.25 14.15v4.25c0 1.094-.787 2.036-1.872 2.18-2.087.277-4.216.42-6.378.42s-4.291-.143-6.378-.42c-1.085-.144-1.872-1.086-1.872-2.18v-4.25m16.5 0a2.18 2.18 0 00.75-1.661V8.706c0-1.081-.768-2.015-1.837-2.175a48.114 48.114 0 00-3.413-.387m4.5 8.006c-.194.165-.42.295-.673.38A23.978 23.978 0 0112 15.75c-2.648 0-5.195-.429-7.577-1.22a2.016 2.016 0 01-.673-.38m0 0A2.18 2.18 0 013 12.489V8.706c0-1.081.768-2.015 1.837-2.175a48.111 48.111 0 013.413-.387m7.5 0V5.25A2.25 2.25 0 0013.5 3h-3a2.25 2.25 0 00-2.25 2.25v.894m7.5 0a48.667 48.667 0 00-7.5 0M12 12.75h.008v.008H12v-.008z"
 												/>
 											</svg>
-											<span> Job Description</span>
-										</div>
+											<span class="hidden hover:underline lg:block">
+												{new URL(job.job_description_url).hostname}
+											</span>
+										</a>
 									{/if}
 									<div class="flex items-center space-x-2">
 										<svg
@@ -138,7 +154,7 @@
 											viewBox="0 0 24 24"
 											stroke-width="1.5"
 											stroke="currentColor"
-											class="hidden h-5 w-5 sm:block"
+											class="hidden h-5 w-5 lg:block"
 										>
 											<path
 												stroke-linecap="round"
@@ -147,7 +163,7 @@
 											/>
 										</svg>
 										<span>
-											{job.emailed_at}
+											{new Date(job.emailed_at).toLocaleDateString()}
 										</span>
 									</div>
 								</div>
